+++ src/routes/+page.svelte
@@ -3,7 +3,7 @@
 
 
   const count = writable(0, { trackerCount: 2 });
-  const { trackers: [prev1] } = count;
+  const { previous: [ prev1 ] } = count;
 
 
   count.subscribe((n, o1, o2) =>